requirements.txt
PyQt5
numpy
Pillow
requests
torch
transformers